feat: workspace reload for new files, TLS removal, logging improvements
New files are now automatically detected by the MSBuild workspace and
picked up on the next tool call — no server restart required. FSW sets
a reloadNeeded flag, and ReloadIfNeeded() rebuilds the workspace under
a write lock.

Also: remove top-level statements from LogViewer and TestHarness,
extract ViewerHtml.cs, show full version+hash in LogViewer, inject
FileLogger into WorkspaceManager, log MSBuild discovery on first
workspace load (not at startup), and improve discovery message wording.
